+++ nengo_spa/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from nengo_spa.main import main
+
+
+sys.exit(main(sys.argv))